requirements.txt
Flask~=2.3.0
Flask-Bootstrap~=3.3.7.1
Flask-Moment~=1.0.5
flask-paginate~=2022.1.8
influxdb~=5.3.1
psutil~=5.9.5
python-dotenv~=1.0.0
PyYAML~=6.0
kafka-python==2.0.2
prometheus-api-client==0.5.5
paho-mqtt==2.1.0